Move agent registration logic into the agent file
