+++ sim/core/sim.go
@@ -3,7 +3,6 @@ package core
 import (
 	"fmt"
 	"math/rand"
-	"reflect"
 	"sort"
 	"strings"
 	"time"
@@ -17,31 +16,6 @@ func debugFunc(sim *Simulation) func(string, ...interface{}) {
 	return func(s string, vals ...interface{}) {
 		fmt.Printf("[%0.1f] "+s, append([]interface{}{sim.CurrentTime.Seconds()}, vals...)...)
 	}
-}
-
-type AgentFactory func(*Simulation, *Character, *api.PlayerOptions) Agent
-
-var agentFactories map[string]AgentFactory = make(map[string]AgentFactory)
-
-func RegisterAgentFactory(emptyOptions interface{}, factory AgentFactory) {
-	typeName := reflect.TypeOf(emptyOptions).Name()
-	if _, ok := agentFactories[typeName]; ok {
-		panic("Aleady registered agent factory: " + typeName)
-	}
-	//fmt.Printf("Registering type: %s", typeName)
-
-	agentFactories[typeName] = factory
-}
-
-func makeAgent(sim *Simulation, character *Character, playerOptions *api.PlayerOptions) Agent {
-	typeName := reflect.TypeOf(playerOptions.GetSpec()).Elem().Name()
-
-	factory, ok := agentFactories[typeName]
-	if !ok {
-		panic("No agent factory for type: " + typeName)
-	}
-
-	return factory(sim, character, playerOptions)
 }
 
 type Options struct {
@@ -131,7 +105,7 @@ func NewIndividualSim(params IndividualParams) *Simulation {
 	sim.IndividualParams = params
 
 	character := NewCharacter(params.Equip, params.Race, params.Consumes, params.CustomStats)
-	agent := makeAgent(sim, character, params.PlayerOptions)
+	agent := NewAgent(sim, character, params.PlayerOptions)
 	raid.AddPlayer(agent)
 	raid.AddPlayerBuffs()
